ui-tests issue#3709
